lexpredict_openedgar/requirements/full.txt
alabaster==0.7.10
amqp==1.4.9
anyjson==0.3.3
apipkg==1.4
argon2-cffi==18.1.0
astroid==1.6.4
attrs==18.1.0
awesome-slugify==1.6.5
Babel==2.5.3
backcall==0.1.0
billiard==3.3.0.23
bleach==2.1.3
boto3==1.6.4
botocore==1.9.4
celery==3.1.25
certifi==2018.4.16
chardet==3.0.4
coverage==4.5.1
coveralls==1.3.0
datefinder==0.6.1
decorator==4.3.0
defusedxml==0.5.0
Django==2.2
django-allauth==0.35.0
django-coverage-plugin==1.5.0
django-crispy-forms==1.7.0
django-debug-toolbar==1.9.1
django-environ==0.4.4
django-extensions==2.0.0
django-model-utils==3.1.1
django-redis==4.8.0
django-test-plus==1.0.22
djangorestframework
docutils==0.14
entrypoints==0.2.3
execnet==1.5.0
factory-boy==2.10.0
Faker==0.8.14
flake8==3.5.0
flower==0.9.2
html5lib==1.0.1
idna==2.6
imagesize==1.0.0
ipdb==0.11
ipykernel==4.8.2
ipywidgets==7.2.1
isort==4.3.4
jedi==0.12.0
Jinja2==2.10.1
jmespath==0.9.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
kombu==3.0.37
lazy-object-proxy==1.3.1
lxml==4.5.2
MarkupSafe==1.0
mccabe==0.6.1
mistune==0.8.3
more-itertools==4.1.0
nbconvert==5.3.1
nbformat==4.4.0
nltk==3.2.4
nose==1.3.7
notebook==5.7.1
num2words==0.5.5
numpy==1.16.0
oauthlib==2.0.7
packaging==17.1
pandas==1.1.1
pandocfilters==1.4.2
parso==0.2.0
pexpect==4.5.0
pickleshare==0.7.4
pluggy==0.6.0
prompt-toolkit==1.0.15
psycopg2==2.8
ptyprocess==0.5.2
py==1.5.3
pycodestyle==2.3.1
pycparser==2.18
pyflakes==1.6.0
Pygments==2.2.0
pylint==1.9.0
pylint-celery==0.3
pylint-common==0.2.5
pylint-django==0.11
pylint-plugin-utils==0.2.6
pyparsing==2.2.0
pytest==3.5.1
pytest-cov==2.5.1
pytest-django==3.1.2
pytest-forked==0.2
pytest-pylint==0.9.0
pytest-sugar==0.9.1
pytest-xdist==1.22.2
python-dateutil==2.7.3
python3-openid==3.1.0
pytz==2018.3
pyzmq==17.0.0
qtconsole==4.3.1
redis==2.10.6
regex==2017.9.23
reporters-db==1.0.12.1
requests==2.20.0
requests-oauthlib==0.8.0
s3transfer==0.1.13
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.7.0
sphinxcontrib-websupport==1.0.1
sqlparse==0.2.4
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
text-unidecode==1.2
tika==1.16
tornado==5.0.2
traitlets==4.3.2
typing==3.6.2
Unidecode==0.4.21
urllib3==1.23
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
whitenoise==3.3.1
widgetsnbextension==3.2.1
wrapt==1.10.11
spacy==2.3.2
django-cors-headers==3.5.0